+++ structurizr-dsl/src/main/java/com/structurizr/dsl/Features.java
@@ -1,7 +1,4 @@
 package com.structurizr.dsl;
 
 public final class Features extends com.structurizr.util.Features {
-
-    public static final String ARCHETYPES = "structurizr.feature.dsl.archetypes";
-
 }