+++ structurizr-dsl/src/test/java/com/structurizr/dsl/DslTests.java
@@ -1459,23 +1459,9 @@ workspace extends source-parent.dsl {
     }
 
     @Test
-    void test_archetypes_WhenDisabled() throws Exception {
-        try {
-            File parentDslFile = new File("src/test/resources/dsl/archetypes.dsl");
-            StructurizrDslParser parser = new StructurizrDslParser();
-            parser.parse(parentDslFile);
-            fail();
-        } catch (StructurizrDslParserException e) {
-            assertTrue(e.getMessage().startsWith("Unexpected tokens (expected: !identifiers, group, person, softwareSystem, deploymentEnvironment, element, ->) at line 4"));
-            assertTrue(e.getMessage().endsWith("archetypes {"));
-        }
-    }
-
-    @Test
-    void test_archetypes_WhenEnabled() throws Exception {
+    void test_archetypes() throws Exception {
         File parentDslFile = new File("src/test/resources/dsl/archetypes.dsl");
         StructurizrDslParser parser = new StructurizrDslParser();
-        parser.getFeatures().enable(Features.ARCHETYPES);
         parser.parse(parentDslFile);
         Workspace workspace = parser.getWorkspace();
 
@@ -1492,7 +1478,6 @@ workspace extends source-parent.dsl {
     void test_archetypesForDefaults() throws Exception {
         File parentDslFile = new File("src/test/resources/dsl/archetypes-for-defaults.dsl");
         StructurizrDslParser parser = new StructurizrDslParser();
-        parser.getFeatures().enable(Features.ARCHETYPES);
         parser.parse(parentDslFile);
         Workspace workspace = parser.getWorkspace();
 
@@ -1520,7 +1505,6 @@ workspace extends source-parent.dsl {
     void test_archetypesForExtension() throws Exception {
         File parentDslFile = new File("src/test/resources/dsl/archetypes-for-extension.dsl");
         StructurizrDslParser parser = new StructurizrDslParser();
-        parser.getFeatures().enable(Features.ARCHETYPES);
         parser.parse(parentDslFile);
         Workspace workspace = parser.getWorkspace();